Iso example
Makes use of iso to cleanly transform the state of a component from a record into a monoid.
Once the state is defined as a monoid, it can be used with prisms and traversals. Also, `mempty` can be used as the initial state when loading the application for the first time.